% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{boxscoresParse}
\alias{boxscoresParse}
\title{Load and parse NBA boxscore information}
\usage{
boxscoresParse(files)
}
\arguments{
\item{files}{a vector with the file names of the XML files to be parsed}
}
\value{
dataframe of boxscore information
}
\description{
Parse sportVU XML file with player boxscores data, which includes traditional

boxscore information (points, rebounds, assists, blocks, etc.) as well

as other unique metrics captured by the sportVU data (seconds possessing

ball, feet run, average speed, touches, passes made, etc.).
}
\keyword{XML}
\keyword{boxscores}
