% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{createBlank_motionChart}
\alias{createBlank_motionChart}
\title{Draw horizontal-oriented full-court of NBA basketball court lines, backboard and hoop}
\usage{
createBlank_motionChart(plot.title)
}
\arguments{
\item{plot.title}{The title of the plot}
}
\value{
ggplot object of an NBA full-court
}
\description{
Using the coordinates output from both the drawCourt_ggplot and

drawBackboard_Basket_ggplot functions, create a ggplot outline of

an NBA court.
}
\keyword{NBA}
\keyword{ggplot2}
