% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{createBlank_ShotChart}
\alias{createBlank_ShotChart}
\title{Draw vertical-oriented half-court of NBA basketball court lines, backboard and hoop}
\usage{
createBlank_ShotChart(plot.title)
}
\arguments{
\item{plot.title}{The title of the plot}
}
\value{
ggplot object of an NBA half-court
}
\description{
Using the coordinates output from both the drawCourt_ggplot and

drawBackboard_Basket_ggplot functions, create a ggplot outline of

an NBA court.
}
\keyword{NBA}
\keyword{ggplot2}
