% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{createMotion_Chart}
\alias{createMotion_Chart}
\title{Draw horizontal-oriented full-court of NBA basketball movement-tracking chart}
\usage{
createMotion_Chart(plot.title, x_loc, y_loc)
}
\arguments{
\item{plot.title}{the title of the plot}

\item{x_loc}{vector with x coordinates of a shooter's position}

\item{y_loc}{vector with  y coordinates of a shooter's position}
}
\value{
ggplot object of an NBA player running around the court
}
\description{
Using the coordinates output from both the drawCourt_ggplot and

drawBackboard_Basket_ggplot functions, first create a ggplot outline

of an NBA court. Then add an additional layer of player coordinates,

plotting the (x,y) location of a player 25x per second, revealing a

unique movement-tracking chart of the player's motions on the court.
}
\keyword{NBA}
\keyword{ggplot2}
