% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{createShot_Chart}
\alias{createShot_Chart}
\title{Draw vertical-oriented full-court of NBA basketball shot chart}
\usage{
createShot_Chart(x_loc, y_loc, makes, attempts, plotTitle)
}
\arguments{
\item{x_loc}{vector with x coordinate of the shooter's location on court}

\item{y_loc}{vector with y coordinate of the shooter's location on court}

\item{makes}{vector with the number of successful attempts at each (x_loc,y_loc)}

\item{attempts}{vector with the number of attempts made at each (x_loc,y_loc)}

\item{plotTitle}{the title of the plot}
}
\value{
ggplot object of an NBA player's shot chart
}
\description{
Using the coordinates output from both the drawCourt_ggplot and

drawBackboard_Basket_ggplot functions, first create a ggplot outline

of an NBA court. Then add an additional layer of geom_point() that

displays a player's color-coded shot percentages from various (x,y)

locations on the court.
}
\examples{
shots <- merge(test_shots, 
test_rosters[, names(test_rosters) \%in\% c("global.player.id", "player", "team.name")],
by = "global.player.id", all.x = TRUE)
shots <- halfcourtData(shots)
player.data = shots[shots$player == "Kobe Bryant", ]
chart.data = generateChart_Data(player.data)
plot.title = 'Kobe Bryant Shot Chart'
createShot_Chart(chart.data$x, chart.data$y, chart.data$player.makes, 
chart.data$player.attempts, plot.title)
}
\keyword{NBA}
\keyword{ggplot2}
