% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShotTable.R
\name{createShotTable}
\alias{createShotTable}
\title{Calculates shot metrics for an individual player-game}
\usage{
createShotTable(shot_data)
}
\arguments{
\item{shot_data}{a dataframe of sportVU data with shooting data}
}
\value{
list of dataframes of data
}
\description{
Takes shot data and returns shooting metrics
}
