% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVelocityTable.R
\name{createVelocityTable}
\alias{createVelocityTable}
\title{Create dataframe with movement metrics for individual player-game}
\usage{
createVelocityTable(data)
}
\arguments{
\item{data}{a sportvu dataframe with x_loc, y_loc and game_clock columns}
}
\value{
a dataframe of movement metrics
}
\description{
Takes sportvu dataframe with x_loc, y_loc, game_clock columns,

and calculates / returns a number of movement metrics.
}
