% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeCourt.R
\name{discretizeCourt}
\alias{discretizeCourt}
\title{Discretize the NBA court for plotting hexagons}
\usage{
discretizeCourt(x_loc, y_loc)
}
\arguments{
\item{x_loc}{the x-coordinates}

\item{y_loc}{the y-coordinates}
}
\value{
dataframe of counts for each coordinate
}
\description{
Discretizes the NBA court into 26x47 points
}
\keyword{discretize}
