% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVelocityTable.R
\name{distance}
\alias{distance}
\title{Calculate distances traveled for 1/25 second intervals}
\usage{
distance(xloc, yloc)
}
\arguments{
\item{xloc}{a vector of x coordinate data for one player}

\item{yloc}{a vector of y coordinate data for one player}
}
\value{
vector of distances traveled
}
\description{
Takes coordinate (x,y) vectors as parameters, and returns

the distances traveled from moment to moment
}
