% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{drawBackboard_Basket_ggplot}
\alias{drawBackboard_Basket_ggplot}
\title{Calculate coordinates of NBA basketball court}
\usage{
drawBackboard_Basket_ggplot()
}
\value{
dataframe of coordinates for NBA backboard and hoop
}
\description{
Using the official specifications of NBA backboard and hoop dimensions,

creates a set of coordinates that can be passed to ggplot2's geom_path()

function to draw an NBA court.
}
\keyword{NBA}
\keyword{ggplot2}
