% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{drawCourt_ggplot}
\alias{drawCourt_ggplot}
\title{Calculate coordinates of NBA basketball court}
\usage{
drawCourt_ggplot()
}
\value{
dataframe of coordinates for NBA court lines
}
\description{
Using the official specifications of NBA court lines, creates a set

of coordinates that can be passed to ggplot2's geom_path() function

to draw an NBA court.
}
\keyword{NBA}
\keyword{ggplot2}
