% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{generateChart_Data}
\alias{generateChart_Data}
\title{Convert shot data into player shot percentages by location}
\usage{
generateChart_Data(player.data)
}
\arguments{
\item{player.data}{a dataframe of shot info filtered by player, etc.}
}
\value{
dataframe with information on the number of attempts and successes at each (x,y) location of the player
}
\description{
The function takes as an argument the shot data filtered by a specific

player and other criteria, and computes the number of attempts and

the percent of those attempts that were successful, at each unique (x,y)

location in the input data
}
\keyword{NBA}
\keyword{chart_data}
