% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{getShinyDataframes}
\alias{getShinyDataframes}
\title{Return the value of the global package variable \code{shiny.dataframes}}
\usage{
getShinyDataframes()
}
\value{
the current value of sportVU package global \code{shiny.dataframes}
}
\description{
Return the value of the global package variable \code{shiny.dataframes}
}
