% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{halfcourtData}
\alias{halfcourtData}
\title{Convert NBA shot coordinates to display on half the court}
\usage{
halfcourtData(shot_data)
}
\arguments{
\item{shot_data}{a formatted shots dataframe of shot information (coords)}
}
\value{
shot dataframe with the new x and y locations
}
\description{
Transforms each (x,y) coordinate in the shot dataframe

to the new (x,y) coordinate it corresponds to with respect to

the half court drawn by createBlank_ShotChart() function.
}
\keyword{shot_data}
