% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{masterParse}
\alias{masterParse}
\title{Load and parse all NBA information}
\usage{
masterParse(directory)
}
\arguments{
\item{directory}{a directory path with XML files}
}
\value{
list of dataframes of data
}
\description{
Parse all sportVU XML files in directory passed
}
\keyword{XML}
\keyword{sportVU}
