% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{momentsParse}
\alias{momentsParse}
\title{Load and parse NBA moments information}
\usage{
momentsParse(files)
}
\arguments{
\item{files}{a vector with the file names of the XML files to be parsed}
}
\value{
dataframe of moments information
}
\description{
Parse sportVU XML file with moments data, which includes data

of all NBA moments (passes, dribbles, shots, steals, etc.). Moments

is a more granular version of the playbyplay parser, mainly due to the

inclusion of passing and dribbling moments.
}
\keyword{XML}
\keyword{moments}
