% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{oncourtParse}
\alias{oncourtParse}
\title{Load and parse NBA oncourt information}
\usage{
oncourtParse(files)
}
\arguments{
\item{files}{a vector with the file names of the XML files to be parsed}
}
\value{
dataframe of oncourt.id information
}
\description{
Parse sportVU XML file with oncourt data, where each unique combination

of 10 players appearing on the court in a single corresponds with an

oncourt.id. Use oncourt.id and game.id to identify a unique combination

of 10 players in a specific game.
}
\keyword{XML}
\keyword{oncourt}
