% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{playbyplayParse}
\alias{playbyplayParse}
\title{Load and parse NBA playbyplay information}
\usage{
playbyplayParse(files)
}
\arguments{
\item{files}{a vector with the file names of the XML files to be parsed}
}
\value{
dataframe of playbyplay information
}
\description{
Parse sportVU XML file with playbyplay data, which includes data

of all NBA plays (shots, rebounds, steals, time out, end of quarter,

etc.) for each NBA game parsed.
}
\keyword{XML}
\keyword{playbyplay}
