% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{possessionsParse}
\alias{possessionsParse}
\title{Load and parse NBA possessions information}
\usage{
possessionsParse(files)
}
\arguments{
\item{files}{a vector with the file names of the XML files to be parsed}
}
\value{
dataframe of possessions information
}
\description{
Parse sportVU XML file with possession data, which includes data

for all NBA possessions (points scored, length of possession, touches,

passes, etc.). Less granular than both the playbyplay and moments

parsers.
}
\keyword{XML}
\keyword{possessions}
