% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{resetOptions}
\alias{resetOptions}
\title{Reset the global package variable \code{shiny.dataframes}}
\usage{
resetOptions()
}
\value{
the default value value of sportVU package global \code{shiny.dataframes}
}
\description{
Reset the global package variable \code{shiny.dataframes}
}
