% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{rosterParse}
\alias{rosterParse}
\title{Load and parse NBA roster information}
\usage{
rosterParse(file)
}
\arguments{
\item{file}{the file name of the XML document to be parsed}
}
\value{
dataframe of NBA roster information
}
\description{
Parse sportVU XML file with NBA roster data into dataframe
}
\keyword{XML}
\keyword{rosters,}
