% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionCreator.R
\name{selectionCreatorMovements}
\alias{selectionCreatorMovements}
\title{Generate selectInput dropdown options for ShinyApp Movement Chart}
\usage{
selectionCreatorMovements(df)
}
\arguments{
\item{df}{the shots dataframe from which dropdown options are pulled from}
}
\value{
list of 2 vectors
}
\description{
Generate selectInput dropdown options for ShinyApp
}
