% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionCreator.R
\name{selectionCreatorShots}
\alias{selectionCreatorShots}
\title{Generate selectInput dropdown options for ShinyApp Shot Charts}
\usage{
selectionCreatorShots(df)
}
\arguments{
\item{df}{the shots dataframe from which dropdown options are pulled from}
}
\value{
list of 4 vectors
}
\description{
Generate selectInput dropdown options for ShinyApp
}
