% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setShots}
\alias{setShots}
\title{Set the package's shots dataframe}
\usage{
setShots(df)
}
\arguments{
\item{df}{the new, parsed dataframe to assign}
}
\value{
the changed value of the package global \code{shiny.dataframes}
}
\description{
Set the package's shots dataframe
}
