% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{shotsParse}
\alias{shotsParse}
\title{Load and parse NBA shot information}
\usage{
shotsParse(files)
}
\arguments{
\item{files}{a vector with the file names of the XML files to be parsed}
}
\value{
dataframe of shot information
}
\description{
Parse sportVU XML file with NBA shot data, which includes data

of all NBA shots (coordinates, make/miss, dribbles, etc.). To

be used for shot-chart application.
}
\keyword{XML}
\keyword{shots}
