% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportVU.R
\docType{package}
\name{sportVU}
\alias{sportVU}
\alias{sportVU-package}
\title{A package with built-in ShinyApp for visualizing NBA sports data.}
\description{
This package is useful to users who want to visualize and analyse NBA sports data.
It provides a list of parser functions that the user can call using \code{link{masterParse}} 
function to parse a variety of XML files. After having parsed the shots data and sportvu data,
the user can launch the ShinyApp by calling \code{link{runShotChart}} function. The ShinyApp
helps visuzlize such data through a motion chart and a shot chart.
}
\references{
Hadley Wickham's book at \url{http://r-pkgs.had.co.nz/}
}
\seealso{
\code{link{masterParse}}

\code{link{runShotChart}}
}
