% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{sportvuParse}
\alias{sportvuParse}
\title{Load and parse NBA sportVU information}
\usage{
sportvuParse(files)
}
\arguments{
\item{files}{a vector with the file names of the XML files to be parsed}
}
\value{
dataframe of sportVU information
}
\description{
Parse sportVU XML file with moments data, which includes motion tracking

coordinates of all 10 players as well as the ball, each 25x per second.

This is the most granular parser due to the high frequency of coordinate

data.
}
\keyword{XML}
\keyword{moments}
