% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{test_rosters}
\alias{test_rosters}
\title{A dataframe with general information on players.}
\description{
The rosters data provides information for each player on their global player id, their team, 
their position on the court etc.
}
\keyword{data}
