% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{test_shots}
\alias{test_shots}
\title{A typical shots data dataframe that will be used to create the shots chart in ShinyApp.}
\description{
A shots data dataframe with information on each shooting attempt for each of the games included
in the dataframe. For each shooting attempt, it provides information on the shooter's name, his (x,y)
location in the court at the time of shooting, number of dribbles before the attempt etc.
}
\keyword{data}
