% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{test_sportvu}
\alias{test_sportvu}
\title{A typical motion data dataframe that will be used to create the motion chart.}
\description{
A motion data dataframe with information on the (x,y) location of each player
for each game id.
}
\keyword{data}
