% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVelocityTable.R
\name{time_played}
\alias{time_played}
\title{Calculate time a player played in game}
\usage{
time_played(clock)
}
\arguments{
\item{clock}{a vector of game-clock times}
}
\value{
a vector of times elapsed
}
\description{
Takes vector of game clock times an individual player was on

court, and returns time elapsed by a player on court
}
